+++ Libraries/Opc.Ua.Server/NodeManager/MonitoredItem/MonitoredNode.cs
