+++ Libraries/Opc.Ua.Server/NodeManager/MonitoredItem/MonitoredNodeMonitoredItemManager.cs
@@ -1,0 +1,266 @@
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Opc.Ua.Server
+{
+    /// <summary>
+    /// Manages the montioredItems for a NodeManager
+    /// </summary>
+    public class MonitoredNodeMonitoredItemManager : IMonitoredItemManager
+    {
+        /// <inheritdoc/>
+        public MonitoredNodeMonitoredItemManager(
+            CustomNodeManager2 nodeManager)
+        {
+            m_nodeManager = nodeManager;
+            m_monitoredNodes = new NodeIdDictionary<MonitoredNode2>();
+            m_monitoredItems = new ConcurrentDictionary<uint, IMonitoredItem>();
+        }
+        /// <inheritdoc/>
+        public NodeIdDictionary<MonitoredNode2> MonitoredNodes => m_monitoredNodes;
+        /// <inheritdoc/>
+        public ConcurrentDictionary<uint, IMonitoredItem> MonitoredItems => m_monitoredItems;
+
+        /// <inheritdoc/>
+        public IMonitoredItem CreateMonitoredItem(
+            IServerInternal server,
+            INodeManager nodeManager,
+            ServerSystemContext context,
+            NodeHandle handle,
+            uint subscriptionId,
+            double publishingInterval,
+            DiagnosticsMasks diagnosticsMasks,
+            TimestampsToReturn timestampsToReturn,
+            MonitoredItemCreateRequest itemToCreate,
+            Range euRange,
+            MonitoringFilter filterToUse,
+            double samplingInterval,
+            uint revisedQueueSize,
+            bool createDurable,
+            uint monitoredItemId,
+            Func<ISystemContext, NodeHandle, NodeState, NodeState> AddNodeToComponentCache)
+        {
+            // check if the node is already being monitored.
+            MonitoredNode2 monitoredNode = null;
+
+            if (!m_monitoredNodes.TryGetValue(handle.Node.NodeId, out monitoredNode))
+            {
+                NodeState cachedNode = AddNodeToComponentCache(context, handle, handle.Node);
+                m_monitoredNodes[handle.Node.NodeId] = monitoredNode = new MonitoredNode2(m_nodeManager, cachedNode);
+            }
+
+            handle.Node = monitoredNode.Node;
+            handle.MonitoredNode = monitoredNode;
+
+            // create the item.
+            MonitoredItem datachangeItem = new MonitoredItem(
+                server,
+                m_nodeManager,
+                handle,
+                subscriptionId,
+                monitoredItemId,
+                itemToCreate.ItemToMonitor,
+                diagnosticsMasks,
+                timestampsToReturn,
+                itemToCreate.MonitoringMode,
+                itemToCreate.RequestedParameters.ClientHandle,
+                filterToUse,
+                filterToUse,
+                euRange,
+                samplingInterval,
+                revisedQueueSize,
+                itemToCreate.RequestedParameters.DiscardOldest,
+                0,
+                createDurable);
+
+            // save the monitored item.
+            monitoredNode.Add(datachangeItem);
+            m_monitoredItems.AddOrUpdate(monitoredItemId, datachangeItem, (key, oldValue) => datachangeItem);
+
+
+            return datachangeItem;
+        }
+        /// <inheritdoc/>
+        public void ApplyChanges()
+        {
+            //only needed for sampling groups
+            return;
+        }
+
+        /// <inheritdoc/>
+        public void Dispose()
+        {
+            //only needed for sampling groups
+            return;
+        }
+
+        /// <inheritdoc/>
+        public StatusCode DeleteMonitoredItem(ServerSystemContext context, IMonitoredItem monitoredItem, NodeHandle handle)
+        {
+            // check for valid monitored item.
+            MonitoredItem datachangeItem = monitoredItem as MonitoredItem;
+
+            // check if the node is already being monitored.
+            MonitoredNode2 monitoredNode = null;
+
+            if (m_monitoredNodes.TryGetValue(handle.NodeId, out monitoredNode))
+            {
+                monitoredNode.Remove(datachangeItem);
+
+                // check if node is no longer being monitored.
+                if (!monitoredNode.HasMonitoredItems)
+                {
+                    m_monitoredNodes.Remove(handle.NodeId);
+                }
+            }
+            else
+            {
+                return StatusCodes.BadMonitoredItemIdInvalid;
+            }
+
+            return StatusCodes.Good;
+        }
+
+        /// <inheritdoc/>
+        public (ServiceResult, MonitoringMode?) SetMonitoringMode(ServerSystemContext context, IMonitoredItem monitoredItem, MonitoringMode monitoringMode, NodeHandle handle)
+        {
+            // check for valid monitored item.
+            MonitoredItem datachangeItem = monitoredItem as MonitoredItem;
+
+            // update monitoring mode.
+            MonitoringMode previousMode = datachangeItem.SetMonitoringMode(monitoringMode);
+
+            // must send the latest value after enabling a disabled item.
+            if (monitoringMode == MonitoringMode.Reporting && previousMode == MonitoringMode.Disabled)
+            {
+                handle.MonitoredNode.QueueValue(context, handle.Node, datachangeItem);
+            }
+
+            return (StatusCodes.Good, previousMode);
+        }
+
+        /// <inheritdoc/>
+        public bool RestoreMonitoredItem(IServerInternal server,
+                                         INodeManager nodeManager,
+                                         ServerSystemContext context,
+                                         NodeHandle handle,
+                                         IStoredMonitoredItem storedMonitoredItem,
+                                         IUserIdentity savedOwnerIdentity,
+                                         Func<ISystemContext, NodeHandle, NodeState, NodeState> AddNodeToComponentCache,
+                                         out IMonitoredItem monitoredItem)
+        {
+            // check if the node is already being monitored.
+            MonitoredNode2 monitoredNode = null;
+
+            if (!m_monitoredNodes.TryGetValue(handle.Node.NodeId, out monitoredNode))
+            {
+                NodeState cachedNode = AddNodeToComponentCache(context, handle, handle.Node);
+                m_monitoredNodes[handle.Node.NodeId] = monitoredNode = new MonitoredNode2(m_nodeManager, cachedNode);
+            }
+
+            handle.Node = monitoredNode.Node;
+            handle.MonitoredNode = monitoredNode;
+
+            // create the item.
+            var datachangeItem = new MonitoredItem(
+                server,
+                nodeManager,
+                handle,
+                storedMonitoredItem);
+
+            // update monitored item list.
+            monitoredItem = datachangeItem;
+
+            // save the monitored item.
+            monitoredNode.Add(datachangeItem);
+
+            return true;
+        }
+        /// <inheritdoc/>
+        public ServiceResult ModifyMonitoredItem(ServerSystemContext context,
+                                                 DiagnosticsMasks diagnosticsMasks,
+                                                 TimestampsToReturn timestampsToReturn,
+                                                 MonitoringFilter filterToUse,
+                                                 Range euRange,
+                                                 double samplingInterval,
+                                                 uint revisedQueueSize,
+                                                 ISampledDataChangeMonitoredItem monitoredItem,
+                                                 MonitoredItemModifyRequest itemToModify)
+        {
+            // modify the monitored item parameters.
+            return monitoredItem.ModifyAttributes(
+                diagnosticsMasks,
+                timestampsToReturn,
+                itemToModify.RequestedParameters.ClientHandle,
+                filterToUse,
+                filterToUse,
+                euRange,
+                samplingInterval,
+                revisedQueueSize,
+                itemToModify.RequestedParameters.DiscardOldest);
+        }
+        /// <inheritdoc/>
+        public (MonitoredNode2, ServiceResult) SubscribeToEvents(ServerSystemContext context, NodeState source, IEventMonitoredItem monitoredItem, bool unsubscribe)
+        {
+            MonitoredNode2 monitoredNode = null;
+            // handle unsubscribe.
+            if (unsubscribe)
+            {
+                // check for existing monitored node.
+                if (!m_monitoredNodes.TryGetValue(source.NodeId, out monitoredNode))
+                {
+                    return (null, StatusCodes.BadNodeIdUnknown);
+                }
+
+                monitoredNode.Remove(monitoredItem);
+                m_monitoredItems.TryRemove(monitoredItem.Id, out _);
+
+                // check if node is no longer being monitored.
+                if (!monitoredNode.HasMonitoredItems)
+                {
+                    m_monitoredNodes.Remove(source.NodeId);
+                }
+
+                return (monitoredNode, ServiceResult.Good);
+            }
+
+            // only objects or views can be subscribed to.
+            if (!(source is BaseObjectState instance) || (instance.EventNotifier & EventNotifiers.SubscribeToEvents) == 0)
+            {
+                if (!(source is ViewState view) || (view.EventNotifier & EventNotifiers.SubscribeToEvents) == 0)
+                {
+                    return (null, StatusCodes.BadNotSupported);
+                }
+            }
+
+            // check for existing monitored node.
+            if (!m_monitoredNodes.TryGetValue(source.NodeId, out monitoredNode))
+            {
+                m_monitoredNodes[source.NodeId] = monitoredNode = new MonitoredNode2(m_nodeManager, source);
+            }
+
+            if (monitoredNode.EventMonitoredItems != null)
+            {
+                // remove existing monitored items with the same Id prior to insertion in order to avoid duplicates
+                // this is necessary since the SubscribeToEvents method is called also from ModifyMonitoredItemsForEvents
+                monitoredNode.EventMonitoredItems.RemoveAll(e => e.Id == monitoredItem.Id);
+            }
+
+            // this links the node to specified monitored item and ensures all events
+            // reported by the node are added to the monitored item's queue.
+            monitoredNode.Add(monitoredItem);
+            m_monitoredItems.TryAdd(monitoredItem.Id, monitoredItem);
+
+            return (monitoredNode, ServiceResult.Good);
+        }
+
+        private readonly CustomNodeManager2 m_nodeManager;
+        private readonly NodeIdDictionary<MonitoredNode2> m_monitoredNodes;
+        private readonly ConcurrentDictionary<uint, IMonitoredItem> m_monitoredItems;
+    }
+
+}
